refactor(core): extract grid child legend construction
